backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
anthropic
supabase
python-dateutil